requirements.txt
Flask==2.2.5
gunicorn==23.0.0
Werkzeug==3.0.6
requests>=2.28.0
Pillow>=9.0.0
openai>=1.0.0
stability-sdk>=0.8.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
line-bot-sdk>=3.0.0
facebook-sdk>=3.1.0
google-generativeai>=0.3.0